Documenting using Chilipeppr
First attempt at documenting use of Chilipeppr to engrave a PCB from and Eagle design on a cheap chinese no-name CNC mill/engraver.
The GBRL 1.1h variation on the wiring for Z-Limit switch changes, Spindle Speed and the use of the Probe input is covered in detail. This varies considerably in small, but important aspects from Tiny-G systems.
The use of the Probe feature in GBRL is described in detail for zeroing the Z-Axis and the functions of Auto Level and Tool Changes is included.
The reduction of RF causing false Alarm:1 triggers is also covered for reliable Auto Level probing.